refactor(install): extract shared advisory install lock
P5 slice 3. The agent-package installer has an advisory install lock (pid +
liveness, stale-claim, idempotent release); the plugin install path has none —
a real concurrency gap. Lift the mechanism into the shared install core,
path-parameterized, so Phase 6 can give plugin install the same guarantee.

- Add src/core/install-core/install-lock.ts: acquireLock/releaseLock/isLockHeld
  keyed by an arbitrary lock path.
- src/core/agent-packages/install-lock.ts becomes a thin wrapper preserving the
  exact ~/.bakin/packages/.lock path and public API.

Behavior-preserving: typecheck clean; characterization (lock release + mutual
exclusion) + installer suites green.
